Adds a config file
I made this earlier, but my antivirus deleted that version. It's very possible that this one has a few bugs. This config file makes it easier to use the hack, but it has a few bad sides like that the config has to be manually pre configured before launching the cheat. 

Also you should consider making a released version where you put everything in one file and compile it. This will remove the requirement to have ahk installed.